Checking now based on phylonium and new blast
So far, the program worked as follows:

1) Use macle to look for for regions in target representative that do
not appear in the neighbors.

2) Regions from 1) with exact matches in all targets are retained.

3) Regions from 2) with no blastn (-task) matches in neighbors are
retained.

This lead to the loss of many potentially valuable regions. To avoid
this, the program now has revised steps 2) and 3) while step 1) is
unchanged.

New 2) Write all regions found in 1) to file and use phylonium to
intersect them with all other target sequences. Regions at least 100 bp
long (-l) are picked from the phylonium output and the positions of
mutations marked as N.

New 3) The regions from 2) are blasted (-task blastn) agains the
neighbors. Homologous intervals are set to N. Templates with at least
100 (-l) nucleotides are printed to stdout.

This new strategy leads to a much better yield of template sequences.