Fixed error in second subtraction
I used to summarize the BLAST hits per query assuming that each query
would at most receive one clique of overlapping BLAST hits. This is, of
course, not true, and led to the removal of nucleotides bracketed
between hits. I have fixed this by not bothering about merging BLAST
intervals any more and simply setting each one to N. This may lead to a
bit of redundant work, but it is simple and works.