Reverting versioning
The versioning is now essentially as before, git is interrogated about
the version and this is injected at compile time. Now, if the program is
distributed as a source archive rather than a git repository, the
version tag will be empty. However, versioning does work as expected if
a user either clones the repository or uses the compiled binaries of a
release.