unbias match lengths by skipping terminal mismatch
I believe this is wrong. After a match of say ten bases we know that the eleventh is a mismatch. Hence if we are starting from there we are biasing ourself to a short random match. Instead we need to skip that base and start our next match one base beyond. This is how it is implemented in Phylonium:

https://github.com/EvolBioInf/phylonium/blob/master/src/process.cxx#L280-L281